requirements.txt
apns2==0.7.2
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
Django==3.1.7
django-axes==5.13.1
django-codemirror2==0.2
django-cte==1.1.4
django-ipware==3.0.2
django-js-asset==1.2.2
django-mfa2==2.1.2
django-mptt==0.9.1
django-push-notifications==2.0.0
django-river==3.3.0
django-simple-history==2.12.0
django-tenants==3.2.1
ecdsa==0.14.1
fido2==0.9.1
h2==2.6.2
hpack==3.0.0
http-ece==1.1.0
hyper==0.7.0
hyperframe==3.2.0
idna==2.10
jsonfield==3.1.0
jsonLookup==0.9.0
psycopg2==2.8.6
py-vapid==1.8.0
pyasn1==0.4.8
pycparser==2.20
PyJWT==1.7.1
pyotp==2.6.0
python-jose==3.2.0
python-u2flib-server==5.0.1
pytz==2021.1
pywebpush==1.13.0
requests==2.25.1
rsa==4.7.2
semantic-version==2.8.5
setuptools-rust==0.12.1
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
ua-parser==0.10.0
urllib3==1.26.4
user-agents==2.2.0
